requirements.txt
# Core LangGraph and LangChain dependencies
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-google-genai>=0.1.0

# OpenAI API client
openai>=1.0.0

# Environment variable management
python-dotenv>=1.0.0

# Type hints support
typing-extensions>=4.0.0

# Optional: For better development experience
pydantic>=2.0.0

# Database and data manipulation
pandas>=2.0.0

# MongoDB support for NoSQL operations
pymongo>=4.6.0
dnspython>=2.4.0

# PostgreSQL support (using binary version for easier installation)
psycopg2-binary>=2.9.0

# Google Generative AI
google-generativeai>=0.1.0

# SQLite3 (built into Python) - Now using PostgreSQL instead
# sqlite3>=2.0.0

# Web framework for deployment
Flask>=2.3.0
gunicorn>=21.0.0
